+++ qt/scientific_interfaces/test/ISISReflectometry/Event/EventPresenterTest.h
@@ -8,7 +8,7 @@
 #define MANTID_CUSTOMINTERFACES_REFLEVENTPRESENTERTEST_H_
 
 #include "../../../ISISReflectometry/GUI/Event/EventPresenter.h"
-#include "../../ReflMockObjects.h"
+#include "../ReflMockObjects.h"
 #include "MockEventView.h"
 #include <cxxtest/TestSuite.h>
 #include <gmock/gmock.h>
